+++ lib/hexapdf/parser.rb
@@ -239,7 +239,7 @@ module HexaPDF
         start.upto(start + number_of_entries - 1) do |oid|
           pos, gen, type = @tokenizer.next_xref_entry do |matched_size|
             maybe_raise("Invalid cross-reference subsection entry", pos: @tokenizer.pos,
-                        force: matched_size == 20)
+                        force: !matched_size)
           end
           if xref.entry?(oid)
             next